Move join models to the separate folder
