+++ src/ggrc-client/js/models/join-models/index.js
@@ -1,0 +1,12 @@
+/*
+  Copyright (C) 2018 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import './join';
+import './multitype-search-join';
+import './object-person';
+import './relationship';
+import './snapshot';
+import './task-group-object';
+import './user-role';
