+++ src/ggrc-client/js/models/join-models/multitype-search-join.js
@@ -1,0 +1,10 @@
+/*
+    Copyright (C) 2018 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import Join from './join';
+
+export default Join('CMS.Models.MultitypeSearchJoin', {
+  join_keys: {},
+}, {});
