+++ src/ggrc-client/js/models/join-models/object-person.js
@@ -1,0 +1,25 @@
+/*
+    Copyright (C) 2018 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import Join from './join';
+
+export default Join('CMS.Models.ObjectPerson', {
+  root_object: 'object_person',
+  root_collection: 'object_people',
+  findAll: 'GET /api/object_people',
+  create: 'POST /api/object_people',
+  update: 'PUT /api/object_people/{id}',
+  destroy: 'DELETE /api/object_people/{id}',
+  join_keys: {
+    personable: can.Model.Cacheable,
+    person: CMS.Models.Person,
+  },
+  attributes: {
+    context: 'CMS.Models.Context.stub',
+    modified_by: 'CMS.Models.Person.stub',
+    person: 'CMS.Models.Person.stub',
+    personable: 'CMS.Models.get_stub',
+  },
+}, {});
